ggsurvplot Argument Descriptions

ggsurvplot Argument Descriptions

Arguments

fit

an object of class survfit.

data

a dataset used to fit survival curves. If not supplied then data will be extracted from 'fit' object.

fun

an arbitrary function defining a transformation of the survival curve. Often used transformations can be specified with a character argument: "event" plots cumulative events (f(y) = 1-y), "cumhaz" plots the cumulative hazard function (f(y) = -log(y)), and "pct" for survival probability in percentage.

surv.scale

scale transformation of survival curves. Allowed values are "default" or "percent".

xscale

numeric or character value specifying x-axis scale.

If numeric, the value is used to divide the labels on the x axis. For example, a value of 365.25 will give labels in years instead of the original days.

If character, allowed options include one of c("d_m", "d_y", "m_d", "m_y", "y_d", "y_m"), where d = days, m = months and y = years. For example, xscale = "d_m" will transform labels from days to months; xscale = "m_y", will transform labels from months to years.

color

color to be used for the survival curves.

If the number of strata/group (n.strata) = 1, the expected value is the color name. For example color = "blue".

If n.strata > 1, the expected value is the grouping variable name. By default, survival curves are colored by strata using the argument color = "strata", but you can also color survival curves by any other grouping variables used to fit the survival curves. In this case, it's possible to specify a custom color palette by using the argument palette.

palette

the color palette to be used. Allowed values include "hue" for the default hue color scale; "grey" for grey color palettes; brewer palettes e.g. "RdBu", "Blues", ...; or custom color palette e.g. c("blue", "red"). See details section for more information. Can be also a numeric vector of length(groups); in this case a basic color palette is created using the function palette.

linetype

line types. Allowed values includes i) "strata" for changing linetypes by strata (i.e. groups); ii) a numeric vector (e.g., c(1, 2)) or a character vector c("solid", "dashed").

break.time.by

numeric value controlling time axis breaks. Default value is NULL.

break.x.by

alias of break.time.by. Numeric value controlling x axis breaks. Default value is NULL.

break.y.by

same as break.x.by but for y axis.

conf.int

logical value. If TRUE, plots confidence interval.

conf.int.fill

fill color to be used for confidence interval.

conf.int.style

confidence interval style. Allowed values include c("ribbon", "step").

conf.int.alpha

numeric value specifying fill color transparency. Value should be in [0, 1], where 0 is full transparency and 1 is no transparency.

censor

logical value. If TRUE, censors will be drawn.

censor.shape

character or numeric value specifying the point shape of censors. Default value is "+" (3), a sensible choice is "|" (124).

censor.size

numveric value specifying the point size of censors. Default is 4.5.

pval

logical value, a numeric or a string. If logical and TRUE, the p-value is added on the plot. If numeric, than the computet p-value is substituted with the one passed with this parameter. If character, then the customized string appears on the plot. See examples - Example 3.

pval.size

numeric value specifying the p-value text size. Default is 5.

pval.coord

numeric vector, of length 2, specifying the x and y coordinates of the p-value. Default values are NULL.

title, xlab, ylab

main title and axis labels

xlim, ylim

x and y axis limits e.g. xlim = c(0, 1000), ylim = c(0, 1).

axes.offset

logical value. Default is TRUE. If FALSE, set the plot axes to start at the origin.

legend

character specifying legend position. Allowed values are one of c("top", "bottom", "left", "right", "none"). Default is "top" side position. to remove the legend use legend = "none". Legend position can be also specified using a numeric vector c(x, y); see details section.

legend.title

legend title.

legend.labs

character vector specifying legend labels. Used to replace the names of the strata from the fit. Should be given in the same order as those strata.

risk.table

Allowed values include:

TRUE or FALSE specifying whether to show or not the risk table. Default is FALSE.

"absolute" or "percentage". Shows the absolute number and the percentage of subjects at risk by time, respectively.

"abs_pct" to show both absolute number and percentage.

"nrisk_cumcensor" and "nrisk_cumevents". Show the number at risk and, the cumulative number of censoring and events, respectively.

risk.table.title

The title to be used for the risk table.

risk.table.pos

character vector specifying the risk table position. Allowed options are one of c("out", "in") indicating 'outside' or 'inside' the main plot, respectively. Default value is "out".

risk.table.col

same as tables.col but for risk table only.

risk.table.fontsize, fontsize

font size to be used for the risk table and the cumulative events table.

risk.table.y.text

logical. Default is TRUE. If FALSE, risk table y axis tick labels will be hidden.

risk.table.y.text.col

logical. Default value is FALSE. If TRUE, risk table tick labels will be colored by strata.

tables.height

numeric value (in [0 - 1]) specifying the general height of all tables under the main survival plot.

tables.y.text

logical. Default is TRUE. If FALSE, the y axis tick labels of tables will be hidden.

tables.y.text.col

logical. Default value is FALSE. If TRUE, tables tick labels will be colored by strata.

tables.col

color to be used for all tables under the main plot. Default value is "black". If you want to color by strata (i.e. groups), use tables.col = "strata".

tables.theme

function, ggplot2 theme name. Default value is theme_survminer. Allowed values include ggplot2 official themes: see theme.

risk.table.height

the height of the risk table on the grid. Increase the value when you have many strata. Default is 0.25. Ignored when risk.table = FALSE.

surv.plot.height

the height of the survival plot on the grid. Default is 0.75. Ignored when risk.table = FALSE.

ncensor.plot

logical value. If TRUE, the number of censored subjects at time t is plotted. Default is FALSE. Ignored when cumcensor = TRUE.

ncensor.plot.title

The title to be used for the censor plot. Used when ncensor.plot = TRUE.

ncensor.plot.height

The height of the censor plot. Used when ncensor.plot = TRUE.

cumevents

logical value specifying whether to show or not the table of the cumulative number of events. Default is FALSE.

cumevents.title

The title to be used for the cumulative events table.

cumevents.col

same as tables.col but for the cumulative events table only.

cumevents.y.text

logical. Default is TRUE. If FALSE, the y axis tick labels of the cumulative events table will be hidden.

cumevents.y.text.col

logical. Default value is FALSE. If TRUE, the y tick labels of the cumulative events will be colored by strata.

cumevents.height

the height of the cumulative events table on the grid. Default is 0.25. Ignored when cumevents = FALSE.

cumcensor

logical value specifying whether to show or not the table of the cumulative number of censoring. Default is FALSE.

cumcensor.title

The title to be used for the cumcensor table.

cumcensor.col

same as tables.col but for cumcensor table only.

cumcensor.y.text

logical. Default is TRUE. If FALSE, the y axis tick labels of the cumcensor table will be hidden.

cumcensor.y.text.col

logical. Default value is FALSE. If TRUE, the y tick labels of the cumcensor will be colored by strata.

cumcensor.height

the height of the cumcensor table on the grid. Default is 0.25. Ignored when cumcensor = FALSE.

surv.median.line

character vector for drawing a horizontal/vertical line at median survival. Allowed values include one of c("none", "hv", "h", "v"). v: vertical, h:horizontal.

ggtheme

function, ggplot2 theme name. Default value is theme_survminer. Allowed values include ggplot2 official themes: see theme.

...

other arguments to be passed i) to ggplot2 geom_*() functions such as linetype, size, ii) or to the function ggpar() for customizing the plots. See details section.

log.rank.weights

The name for the type of weights to be used in computing the p-value for log-rank test. By default survdiff is used to calculate regular log-rank test (with weights == 1). A user can specify "1", "n", "sqrtN", "S1", "S2", "FH" to use weights specified in comp, so that weight correspond to the test as : 1 - log-rank, n - Gehan-Breslow (generalized Wilcoxon), sqrtN - Tarone-Ware, S1 - Peto-Peto's modified survival estimate, S2 - modified Peto-Peto (by Andersen), FH - Fleming-Harrington(p=1, q=1).

pval.method

whether to add a text with the test name used for calculating the pvalue, that corresponds to survival curves' comparison - used only when pval=TRUE

pval.method.size

the same as pval.size but for displaying log.rank.weights name

pval.method.coord

the same as pval.coord but for displaying log.rank.weights name