Multiple Myeloma Data

Multiple Myeloma data extracted from publicly available gene expression data (GEO Id: GSE4581).

data("myeloma")

Format

A data frame with 256 rows and 12 columns.

molecular_group

Patients' molecular subgroups

chr1q21_status

Amplification status of the chromosome 1q21

treatment

treatment

event

survival status 0 = alive, 1 = dead

time

Survival time in months

CCND1

Gene expression

CRIM1

Gene expression

DEPDC1

Gene expression

IRF4

Gene expression

TP53

Gene expression

WHSC1

Gene expression

The remaining columns (CCND1, CRIM1, DEPDC1, IRF4, TP53, WHSC1) correspond to the gene expression level of specified genes.

Examples

data(myeloma)
head(myeloma)

#>          molecular_group chr1q21_status treatment event  time   CCND1 CRIM1
#> GSM50986      Cyclin D-1       3 copies       TT2     0 69.24  9908.4 420.9
#> GSM50988      Cyclin D-2       2 copies       TT2     0 66.43 16698.8  52.0
#> GSM50989           MMSET       2 copies       TT2     0 66.50   294.5 617.9
#> GSM50990           MMSET       3 copies       TT2     1 42.67   241.9  11.9
#> GSM50991             MAF           <NA>       TT2     0 65.00   472.6  38.8
#> GSM50992    Hyperdiploid       2 copies       TT2     0 65.20   664.1  16.9
#>          DEPDC1    IRF4   TP53   WHSC1
#> GSM50986  523.5 16156.5   10.0   261.9
#> GSM50988   21.1 16946.2 1056.9   363.8
#> GSM50989  192.9  8903.9 1762.8 10042.9
#> GSM50990  184.7 11894.7  946.8  4931.0
#> GSM50991  212.0  7563.1  361.4   165.0
#> GSM50992  341.6 16023.4 2096.3   569.2